+++ NEWS.md
@@ -1,5 +1,6 @@
 # baselinenowcast 0.0.0.1000
 
+-   Add intermediate-level grouping functions to generate a point nowcast from a reporting triangle and to take a point nowcast and estimate and apply uncertainty.
 -   Refactor `estimate_uncertainty` to take in an error model function, an aggregator function for aggregating across reference times, and an aggregator function for aggregating across delays.
 -   Fix internal checks that ensure there is sufficient data for the specified target choice, using the number of NA rows rather than the number of columns as a proxy for the number of horizons.
 -   Replace argument names with more complete versions where possible.